auto-claude: subtask-2-3 - Create usePlayerManagement hook
